moved to separate bounding box module
