+++ tests/evaluation/scoring_methods/bounding_box_intersection_test.py
@@ -5,41 +5,6 @@ from sciencebeam_judge.evaluation.scoring_methods.bounding_box_intersection impo
 )
 
 
-class TestBoundingBox:
-    def test_should_calculate_area(self):
-        bounding_box = BoundingBox(
-            x=101, y=102, width=200, height=50
-        )
-        assert bounding_box.area == 200 * 50
-
-    def test_should_calculate_intersection_with_identical_bounding_box(self):
-        bounding_box = BoundingBox(110, 120, 50, 60)
-        assert (
-            bounding_box.intersection(bounding_box) == bounding_box
-        )
-
-    def test_should_calculate_intersection_with_smaller_contained_bounding_box(self):
-        assert (
-            BoundingBox(100, 100, 200, 200).intersection(
-                BoundingBox(110, 120, 50, 60)
-            ) == BoundingBox(110, 120, 50, 60)
-        )
-
-    def test_should_calculate_intersection_with_larger_bounding_box(self):
-        assert (
-            BoundingBox(110, 120, 50, 60).intersection(
-                BoundingBox(100, 100, 200, 200)
-            ) == BoundingBox(110, 120, 50, 60)
-        )
-
-    def test_should_calculate_intersection_with_overlapping_bounding_box(self):
-        assert (
-            BoundingBox(110, 120, 50, 60).intersection(
-                BoundingBox(120, 110, 100, 100)
-            ) == BoundingBox(120, 120, 40, 60)
-        )
-
-
 class TestParsePageBoundingBoxList:
     def test_should_parse_empty_none(self):
         result = parse_page_bounding_box_list(None)
